requirements.txt
asgiref==3.5.2
black==22.10.0
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==3.2.13
django-appconf==1.0.5
django-bootstrap5==22.1
django-imagekit==4.1.0
django-rest-framework
gunicorn==20.1.0
mypy-extensions==0.4.3
pathspec==0.10.1
pilkit==2.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
whitenoise==6.2.0
